Move some dependencies into the source tree
