+++ src/xbot/common/controls/sensors/mock_adapters/MockJoystick.java
@@ -1,4 +1,4 @@
-package edu.wpi.first.wpilibj;
+package xbot.common.controls.sensors.mock_adapters;
 
 import java.util.HashMap;
 import java.util.Map;
