+++ tests/xbot/common/subsystems/pose/TestPoseSubsystem.java
@@ -3,8 +3,8 @@ package xbot.common.subsystems.pose;
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
 
-import xbot.common.controls.actuators.MockCANTalon;
 import xbot.common.controls.actuators.XCANTalon;
+import xbot.common.controls.actuators.mock_adapters.MockCANTalon;
 import xbot.common.injection.wpi_factories.CommonLibFactory;
 import xbot.common.properties.XPropertyManager;
 import xbot.common.subsystems.pose.BasePoseSubsystem;
